Print manifest for Unlocked Package
