+++ packages/azpipelines/BuildTasks/InstallUnlockedPackageTask/InstallUnlockedPackage.ts
@@ -99,7 +99,8 @@ async function run() {
       options,
       wait_time,
       publish_wait_time,
-      skip_if_package_installed
+      skip_if_package_installed,
+      packageMetadataFromStorage
     );
     let endTime=Date.now()-startTime;
